+++ types.js
@@ -87,8 +87,5 @@ export type Context = {
   query: { [string]: string },
   err?: Error | { statusCode: number },
   res?: $Response,
-  req?: $Request,
-  // The accessToken isn't really part of the context object passed to
-  accessToken: string,
-  isAuthenticated: boolean
+  req?: $Request
 };
